moved the camera to the game class, and also changed how it takes input.
really like how movement input is checked now, there is a new function called getMovement that gets wasd and returns a vector, shrinking it if movement is diagonal. So, all that the camera has to do is get the movement and apply it with the speed.

also some simple conversions are now in a file called conversions to make my life easier