collision detection with collision attribute and collision rects
the collision rects have functions that get the borders to make it easy to get important positions, and they have the game position so they can directly set their entity's position

i changed the tps to 60, and for some reason it can't hit that exactly, it gets 61-63, even though it could do 40 exactly. Not sure what's up with that but it's a tiny issue.

A lot of progress today, next step is resolving the collision properly.