collision rects can now be offset and have a different size from their sprite
you no longer have to make variables for textures, just calling the getTexture() function will load them into the map if they havent been yet. This works for me but maybe this would be bad for a bigger game with bigger or more textures, since it could cause a stutter.

a bush has been added.